Separate consume logic from assignment logic
This fixes an issue where failure to start consuming on a partition
causes unnecessary rebalance loops.  We had an issue where a
broker failed, so attempts to subscribe were failing with
ErrLeaderNotAvailable.  Previously, that would fail the rebalance,
which would trigger another rebalance ad finitum.

Instead of inlining the subscribe to the rebalance, this change
sends it to the background and allows it to self-heal without
affecting other consumers.